app-rutas/requirements copy.txt
Flask>2.0
numpy>1.0
networkx>3.0
matplotlib>3.0
Pillow>9.0
mysql-connector-python>8.0
flask-swagger-ui>3.0
Flask-Cors>4.0